refactor(managecategories): rename setupcategories to managecategories
